Lineage patch 3: add customized type LineageArrayType to expose lineage array to the Python interface using function Individual.lineage (not defined yet). Python 3 support is pending.

+++ src/genoStru.cpp
@@ -844,7 +844,7 @@ pairu GenoStruTrait::chromLocusPair(size_t locus) const
 }
 
 
-string GenoStruTrait::alleleName(const UINT allele, const size_t locus) const
+string GenoStruTrait::alleleName(const ULONG allele, const size_t locus) const
 {
 	DBG_FAILIF(allele > ModuleMaxAllele,
 		IndexError, "Allele " + toStr(allele) + " out of range of 0 ~ " +
